ChatBotBackend/requirements.txt
flask
flask-cors
python-dotenv
psycopg2
werkzeug
PyJWT
transformers
langchain
chromadb
sentence-transformers
ollama
langchain-chroma
langchain-ollama